DDF-576: Create an anonymous handler
The UI has been changed to support the updated security handler. Changes
include a new error message displayed above the username label upon attempting
to sign in with invalid credentials, a fix to incorrect/improper cursor when
hovering over the username and password labels, and CSS and jQuery updates to
support the updated security handler.