In the Catalog Federation Strategy dialog, it is necessary to disable caching of products as a workaround to the problem described in https://tools.codice.org/jira/browse/DDF-1029.
This solution involves reverting to the following changes:
https://github.com/codice/ddf-catalog/commit/873f7c74520fc097bf764a1e9554046de8904b71

In addition, more logic is required in the CachingFederationStrategy.java to guard against adding of records to the metacard cache when caching is disabled.

The ddf-ui SearchController.java class will also require an update to expect all results from a single 'index' mode in lieu of separately querying a 'cache' mode. When caching is enabled, all results are returned when querying 'index' mode.